pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prefvlm"
version = "0.1.0"
description = "PrefVLM-MVP: Adapting PrefDisco benchmark to vision-language models"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.75.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "diskcache>=5.6.3",
    "tenacity>=8.3.0",
    "loguru>=0.7.2",
    "tqdm>=4.66.4",
    "datasets>=2.21.0",
    "pillow>=10.4.0",
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "pyarrow>=16.1.0",
    "scikit-learn>=1.5.1",
    "matplotlib>=3.9.1",
    "seaborn>=0.13.2",
    "rich>=13.7.1",
    "click>=8.1.7",
    "httpx>=0.27.0",
    "tabulate>=0.10.0",
]

[project.scripts]
prefvlm = "prefvlm.run_mvp:cli"

[tool.hatch.build.targets.wheel]
packages = ["prefvlm"]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.0",
    "nbformat>=5.10.0",
]
